Fix update/re-rendering logic
Previously, re-rendering a <Popper /> component manually would not update the Popper instance, as long as the `placement` prop did not change and the render function returned a similar DOM tree (so that `popperNode` remained unchanged). E.g., if the render function returned a larger child, Popper.js would not get a chance to re-position the popper element accordingly.

Fix this by always recreating/updating the Popper instance in `componentDidUpdate`, and being somewhat more clever in the `updateStateModifier()` function (just updating the state and hence re-rendering if something of relevance did actually change).